# cmd: fix race in fsDeleteFile, make isDirEmpty deterministic, drop stat in deleteFile, DRY