@brendanashworth 2017-07-31
Running the CI uncovers a porting issue on Windows, looking into this now.