@harshavardhana 2017-07-31
Its better to send three separate PRs here @brendanashworth - they are all unrelated.